+++ app/src/main/java/com/example/tourmanage/ui/components/LoadingWidget.kt
@@ -1,0 +1,34 @@
+package com.example.tourmanage.ui.components
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.CircularProgressIndicator
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.alpha
+import androidx.compose.ui.graphics.Color
+import timber.log.Timber
+
+@Composable
+fun LoadingWidget() {
+    Timber.i("LoadingWidget() is called.")
+    Box(
+        modifier = Modifier
+            .fillMaxSize()
+            .clickable(enabled = false) { },
+        contentAlignment = Alignment.Center
+    ) {
+        Box(
+            modifier = Modifier
+                .fillMaxSize()
+                .alpha(0.1f)
+                .background(Color.DarkGray)
+        )
+        CircularProgressIndicator(
+            color = Color.Red
+        )
+    }
+}
