+++ app/src/main/java/com/example/tourmanage/ui/staydetail/StayImagePopup.kt
@@ -1,0 +1,8 @@
+package com.example.tourmanage.ui.staydetail
+
+import androidx.compose.runtime.Composable
+import com.bumptech.glide.integration.compose.GlideImage
+
+@Composable
+fun StayImagePopup() {
+}
