+++ app/src/main/res/values/colors.xml
@@ -11,4 +11,5 @@
     <color name="white_smoke">#F5F5F5</color>
     <color name="royal_blue">#4169E1</color>
     <color name="cornflower_blue">#6495ED</color>
+    <color name="lightpink">#FFB6C1</color>
 </resources>
